bot_send_billing_to_slack/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-bigquery==1.27.2
